+++ frontend/front/src/pages/myPage/MyComments.jsx
@@ -6,11 +6,6 @@ import AuthHttp from '../../http/authHttp';
 import Layout from '../../components/layout/Layout';
 import MypageNav from '../../components/mypage/myPageNav';
 import {
-  Button,
-  Arrow,
-  DoubleArrow,
-} from '../../components/comment/CommentList';
-import {
   Wrap,
   PageWrap,
   Text,
@@ -23,8 +18,8 @@ import {
   ContentsText,
   ContentsBox,
   ContentsArrow,
-  Nav,
 } from './MyPosts';
+import Pagination from '../../components/mypage/pagination';
 
 const authHttp = new AuthHttp();
 
@@ -43,7 +38,6 @@ const MyComments = () => {
   const [Comments, setComments] = useState([]);
   const [Page, setPage] = useState([]);
   const [SelectedButton, setSelectedButton] = useState(0);
-  const [Count, setCount] = useState(0);
 
   useEffect(() => {
     onMypage();
@@ -68,7 +62,6 @@ const MyComments = () => {
       const res = await authHttp.getCommentList(userId, SelectedButton);
       console.log(res);
       setComments(res.data.result.commentList);
-      setCount(res.data.result.total);
       const arrayLength = res.data.result.total;
       const newArray = new Array(arrayLength).fill(0).map((_, index) => index);
       setPage(newArray);
@@ -78,37 +71,8 @@ const MyComments = () => {
   };
 
   // 페이지 네이션 함수
-  //넘버 버튼으로 페이지 불러오기
-  const pageList = pageNum => {
-    setSelectedButton(pageNum);
-  };
-
-  //left arrow 버튼으로 페이지 불러오기
-  const leftList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(prev => prev - 1);
-    }
-  };
-
-  //right arrow 버튼으로 페이지 불러오기
-  const rightList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(prev => prev + 1);
-    }
-  };
-
-  //첫 페이지로 이동
-  const firstList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(0);
-    }
-  };
-
-  //마지막 페이지로 이동
-  const lastList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(Page.length - 1);
-    }
+  const handlePagination = buttonValue => {
+    setSelectedButton(buttonValue);
   };
 
   return (
@@ -154,33 +118,12 @@ const MyComments = () => {
                   ))}
                 </ContentsWrap>
 
-                <Nav>
-                  {SelectedButton > 0 && (
-                    <Button onClick={() => firstList()}>
-                      <DoubleArrow url='/image/double-arrow-left.png' />
-                    </Button>
-                  )}
-                  <Button onClick={() => leftList()}>
-                    <Arrow url='/image/arrow-left.png' />
-                  </Button>
-                  {Page.map((page, i) => (
-                    <Button
-                      key={i}
-                      onClick={() => pageList(page)}
-                      aria-current={page === SelectedButton ? 'true' : null}
-                    >
-                      {page + 1}
-                    </Button>
-                  ))}
-                  <Button onClick={() => rightList()}>
-                    <Arrow url='/image/arrow-right.png' />
-                  </Button>
-                  {SelectedButton < Page.length - 1 && (
-                    <Button onClick={() => lastList()}>
-                      <DoubleArrow url='/image/double-arrow-right.png' />
-                    </Button>
-                  )}
-                </Nav>
+                {/* 페이지네이션 */}
+                <Pagination
+                  handlePagination={handlePagination}
+                  Page={Page}
+                  SelectedButton={SelectedButton}
+                />
               </>
             )}
           </PageWrap>
