+++ frontend/front/src/pages/myPage/MyLikes.jsx
@@ -6,11 +6,6 @@ import AuthHttp from '../../http/authHttp';
 import Layout from '../../components/layout/Layout';
 import MypageNav from '../../components/mypage/myPageNav';
 import {
-  Button,
-  Arrow,
-  DoubleArrow,
-} from '../../components/comment/CommentList';
-import {
   Wrap,
   Text,
   BoxWrap,
@@ -22,9 +17,9 @@ import {
   ContentsWrap,
   ContentsText,
   ContentsArrow,
-  Nav,
 } from './MyPosts';
 import Alert from '../../components/atomic/modal/Alert';
+import Pagination from '../../components/mypage/pagination';
 
 const authHttp = new AuthHttp();
 
@@ -43,7 +38,6 @@ const MyLikes = () => {
   const [Likes, setLikes] = useState([]);
   const [Page, setPage] = useState([]);
   const [SelectedButton, setSelectedButton] = useState(0);
-  const [Count, setCount] = useState(0);
   const [IsEdit, setIsEdit] = useState(false);
   const [BoardIdList, setBoardIdList] = useState([]);
   const [IsModal, setIsModal] = useState(false);
@@ -73,7 +67,6 @@ const MyLikes = () => {
       const res = await authHttp.getLikeList(userId, SelectedButton);
       console.log('좋아요한 게시글 리스트', res);
       setLikes(res.data.result.commentList);
-      setCount(res.data.result.total);
       const arrayLength = res.data.result.total;
       const newArray = new Array(arrayLength).fill(0).map((_, index) => index);
       setPage(newArray);
@@ -106,38 +99,8 @@ const MyLikes = () => {
     }
   };
 
-  // 페이지 네이션 함수
-  //넘버 버튼으로 페이지 불러오기
-  const pageList = pageNum => {
-    setSelectedButton(pageNum);
-  };
-
-  //left arrow 버튼으로 페이지 불러오기
-  const leftList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(prev => prev - 1);
-    }
-  };
-
-  //right arrow 버튼으로 페이지 불러오기
-  const rightList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(prev => prev + 1);
-    }
-  };
-
-  //첫 페이지로 이동
-  const firstList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(0);
-    }
-  };
-
-  //마지막 페이지로 이동
-  const lastList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(Page.length - 1);
-    }
+  const handlePagination = buttonValue => {
+    setSelectedButton(buttonValue);
   };
 
   //편집모드 닫기
@@ -278,33 +241,12 @@ const MyLikes = () => {
                   ))}
                 </ContentsWrap>
 
-                <Nav>
-                  {SelectedButton > 0 && (
-                    <Button onClick={() => firstList()}>
-                      <DoubleArrow url='/image/double-arrow-left.png' />
-                    </Button>
-                  )}
-                  <Button onClick={() => leftList()}>
-                    <Arrow url='/image/arrow-left.png' />
-                  </Button>
-                  {Page.map((page, i) => (
-                    <Button
-                      key={i}
-                      onClick={() => pageList(page)}
-                      aria-current={page === SelectedButton ? 'true' : null}
-                    >
-                      {page + 1}
-                    </Button>
-                  ))}
-                  <Button onClick={() => rightList()}>
-                    <Arrow url='/image/arrow-right.png' />
-                  </Button>
-                  {SelectedButton < Page.length - 1 && (
-                    <Button onClick={() => lastList()}>
-                      <DoubleArrow url='/image/double-arrow-right.png' />
-                    </Button>
-                  )}
-                </Nav>
+                {/* 페이지네이션 */}
+                <Pagination
+                  handlePagination={handlePagination}
+                  Page={Page}
+                  SelectedButton={SelectedButton}
+                />
               </>
             )}
           </PageWrap>
