+++ frontend/front/src/pages/myPage/MyPosts.jsx
@@ -5,11 +5,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import AuthHttp from '../../http/authHttp';
 import Layout from '../../components/layout/Layout';
 import MypageNav from '../../components/mypage/myPageNav';
-import {
-  Button,
-  Arrow,
-  DoubleArrow,
-} from '../../components/comment/CommentList';
+import Pagination from '../../components/mypage/pagination';
 
 const authHttp = new AuthHttp();
 
@@ -24,11 +20,9 @@ const MyPosts = () => {
   const navigate = useNavigate();
 
   const [UserInfo, setUserInfo] = useState([]);
-
   const [Posts, setPosts] = useState([]);
   const [Page, setPage] = useState([]);
   const [SelectedButton, setSelectedButton] = useState(0);
-  const [Count, setCount] = useState(0);
 
   useEffect(() => {
     onMypage();
@@ -53,7 +47,6 @@ const MyPosts = () => {
       const res = await authHttp.getPostList(userId, SelectedButton);
       console.log(res);
       setPosts(res.data.result.boardList);
-      setCount(res.data.result.total);
       const arrayLength = res.data.result.total;
       const newArray = new Array(arrayLength).fill(0).map((_, index) => index);
       setPage(newArray);
@@ -62,38 +55,8 @@ const MyPosts = () => {
     }
   };
 
-  // 페이지 네이션 함수
-  //넘버 버튼으로 페이지 불러오기
-  const pageList = pageNum => {
-    setSelectedButton(pageNum);
-  };
-
-  //left arrow 버튼으로 페이지 불러오기
-  const leftList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(prev => prev - 1);
-    }
-  };
-
-  //right arrow 버튼으로 페이지 불러오기
-  const rightList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(prev => prev + 1);
-    }
-  };
-
-  //첫 페이지로 이동
-  const firstList = () => {
-    if (SelectedButton > 0) {
-      setSelectedButton(0);
-    }
-  };
-
-  //마지막 페이지로 이동
-  const lastList = () => {
-    if (SelectedButton < Page.length - 1) {
-      setSelectedButton(Page.length - 1);
-    }
+  const handlePagination = buttonValue => {
+    setSelectedButton(buttonValue);
   };
 
   return (
@@ -138,34 +101,12 @@ const MyPosts = () => {
                     </ContentsBox>
                   ))}
                 </ContentsWrap>
-
-                <Nav>
-                  {SelectedButton > 0 && (
-                    <Button onClick={() => firstList()}>
-                      <DoubleArrow url='/image/double-arrow-left.png' />
-                    </Button>
-                  )}
-                  <Button onClick={() => leftList()}>
-                    <Arrow url='/image/arrow-left.png' />
-                  </Button>
-                  {Page.map((page, i) => (
-                    <Button
-                      key={i}
-                      onClick={() => pageList(page)}
-                      aria-current={page === SelectedButton ? 'true' : null}
-                    >
-                      {page + 1}
-                    </Button>
-                  ))}
-                  <Button onClick={() => rightList()}>
-                    <Arrow url='/image/arrow-right.png' />
-                  </Button>
-                  {SelectedButton < Page.length - 1 && (
-                    <Button onClick={() => lastList()}>
-                      <DoubleArrow url='/image/double-arrow-right.png' />
-                    </Button>
-                  )}
-                </Nav>
+                {/* 페이지네이션 */}
+                <Pagination
+                  handlePagination={handlePagination}
+                  Page={Page}
+                  SelectedButton={SelectedButton}
+                />
               </>
             )}
           </PageWrap>
